pamphlet: Render mobile documents with a 5mm margin at 11pt font
A 5mm margin allows marginally more text to fit onto each page,
essential for when the size of the document is this small. Font
size set to 11pts also for this reason.

Only for the mobile doc though, as the -V on pandoc overrides
the geometry section in the YAML header on the markdown file.